% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matSPACE-package.R
\docType{package}
\name{matSPACE-package}
\alias{matSPACE-package}
\title{matSPACE: Sparse Partial Correlation Estimation for Matrix-Variate Data}
\description{
Fits a sparse network of partial correlations among the columns of
matrix-variate data using an L1-penalized (lasso) SPACE-style shooting
algorithm, with optional per-column reweighting and residual variance
re-estimation across outer iterations. See \code{\link[=space]{space()}} for the main
model-fitting function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kimhyew1/matSPACE}
  \item Report bugs at \url{https://github.com/kimhyew1/matSPACE/issues}
}

}
\author{
\strong{Maintainer}: Hyewon Kim \email{kimhw4126@gmail.com}

Authors:
\itemize{
  \item Hyewon Kim \email{kimhw4126@gmail.com}
  \item Seongoh Park \email{seongohpark6@gmail.com}
}

}
\keyword{internal}
